// apiConnection/parser/models/search-base.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LCT Hack Parser
 * Cian.ru parser for LCT Hack
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Segment } from './segment';
import { Walls } from './walls';
/**
 * 
 * @export
 * @interface SearchBase
 */
export interface SearchBase {
    /**
     * 
     * @type {string}
     * @memberof SearchBase
     */
    queryId: string;
    /**
     * 
     * @type {string}
     * @memberof SearchBase
     */
    subqueryId: string;
    /**
     * 
     * @type {string}
     * @memberof SearchBase
     */
    address: string;
    /**
     * 
     * @type {number}
     * @memberof SearchBase
     */
    rooms: number;
    /**
     * 
     * @type {Segment}
     * @memberof SearchBase
     */
    segment: Segment;
    /**
     * 
     * @type {number}
     * @memberof SearchBase
     */
    floors: number;
    /**
     * 
     * @type {Walls}
     * @memberof SearchBase
     */
    walls: Walls;
    /**
     * 
     * @type {number}
     * @memberof SearchBase
     */
    radius?: number;
}